feat: replace Apility with Auth0
Replace Apility with Auth0 because Apility will be depreacated in near future.
