+++ src/lib/searcher/index.ts
@@ -1,9 +1,9 @@
 export { AbuseIPDB } from "./abuseipdb";
 export { AnyRun } from "./anyrun";
-export { Apility } from "./apility";
 export { Apklab } from "./apklab";
 export { ArchiveOrg } from "./archiveorg";
 export { ArchiveToday } from "./archivetoday";
+export { Auth0 } from "./auth0";
 export { BGPView } from "./bgpview";
 export { BinaryEdge } from "./binaryedge";
 export { BitcoinAbuse } from "./bitcoinabuse";
